Refactor wire out of pipe
