+++ choreographer/_channels/__init__.py
@@ -1,0 +1,9 @@
+from ._errors import BlockWarning, ChannelClosedError, JSONError
+from .pipe import Pipe
+
+__all__ = [
+    "BlockWarning",
+    "ChannelClosedError",
+    "JSONError",
+    "Pipe",
+]
