+++ choreographer/_channels/_errors.py
@@ -1,0 +1,10 @@
+class BlockWarning(UserWarning):
+    pass
+
+
+class ChannelClosedError(IOError):
+    pass
+
+
+class JSONError(RuntimeError):
+    pass
