+++ choreographer/_channels/pipe.py
@@ -4,37 +4,15 @@ import sys
 import warnings
 from threading import Lock
 
-import simplejson
+import channel._wire as wire
+from channel._errors import BlockWarning, ChannelClosedError, JSONError
 
-with_block = bool(sys.version_info[:3] >= (3, 12) or platform.system() != "Windows")
-
-
-class PipeClosedError(IOError):
-    pass
-
-
-class BlockWarning(UserWarning):
-    pass
-
-
-class MultiEncoder(simplejson.JSONEncoder):
-    """Special json encoder for numpy types."""
-
-    def default(self, obj):
-        if hasattr(obj, "dtype") and obj.dtype.kind == "i" and obj.shape == ():
-            return int(obj)
-        elif hasattr(obj, "dtype") and obj.dtype.kind == "f" and obj.shape == ():
-            return float(obj)
-        elif hasattr(obj, "dtype") and obj.shape != ():
-            return obj.tolist()
-        elif hasattr(obj, "isoformat"):
-            return obj.isoformat()
-        return simplejson.JSONEncoder.default(self, obj)
+_with_block = bool(sys.version_info[:3] >= (3, 12) or platform.system() != "Windows")
 
 
 # if we're a pipe we expect these public attributes
 class Pipe:
-    def __init__(self, *, debug=False, json_encoder=MultiEncoder):
+    def __init__(self, *, debug=False):
         # This is where pipe listens (from browser)
         # So pass the write to browser
         self._read_from_browser, self._write_from_browser = list(os.pipe())
@@ -54,45 +32,32 @@ class Pipe:
         # windows-style handles. But let another layer handle that.
 
         self.debug = debug  # should be private
-        self.json_encoder = json_encoder  # should be private
 
         # this is just a convenience to prevent multiple shutdowns
         self.shutdown_lock = Lock()  # should be private
 
-    def serialize(self, obj):
-        message = simplejson.dumps(
-            obj,
-            ensure_ascii=False,
-            ignore_nan=True,
-            cls=self.json_encoder,
-        )
-        return message.encode("utf-8") + b"\0"
-
-    def deserialize(self, message):
-        return simplejson.loads(message)
-
     def write_json(self, obj, debug=None):
         if self.shutdown_lock.locked():
-            raise PipeClosedError
+            raise ChannelClosedError
         if not debug:
             debug = self.debug
         if debug:
             print(f"write_json: {obj}", file=sys.stderr)
-        encoded_message = self.serialize(obj)
+        encoded_message = wire.serialize(obj) + b"\0"
         if debug:
             print(f"write_json: {encoded_message}", file=sys.stderr)
         try:
             os.write(self._write_to_browser, encoded_message)
         except OSError as e:
             self.close()
-            raise PipeClosedError from e
+            raise ChannelClosedError from e
         if debug:
             print("wrote_json.", file=sys.stderr)
 
     def read_jsons(self, *, blocking=True, debug=None):  # noqa: PLR0912, C901 branches, complexity
         if self.shutdown_lock.locked():
-            raise PipeClosedError
-        if not with_block and not blocking:
+            raise ChannelClosedError
+        if not _with_block and not blocking:
             warnings.warn(  # noqa: B028
                 "Windows python version < 3.12 does not support non-blocking",
                 BlockWarning,
@@ -106,11 +71,11 @@ class Pipe:
             )
         jsons = []
         try:
-            if with_block:
+            if _with_block:
                 os.set_blocking(self._read_from_browser, blocking)
         except OSError as e:
             self.close()
-            raise PipeClosedError from e
+            raise ChannelClosedError from e
         try:
             raw_buffer = None  # if we fail in read, we already defined
             raw_buffer = os.read(
@@ -121,10 +86,10 @@ class Pipe:
                 # we seem to need {bye} even if chrome closes NOTE
                 if debug:
                     print("read_jsons pipe was closed, raising", file=sys.stderr)
-                raise PipeClosedError
+                raise ChannelClosedError
             while raw_buffer[-1] != 0:
                 # still not great, return what you have
-                if with_block:
+                if _with_block:
                     os.set_blocking(self._read_from_browser, True)
                 raw_buffer += os.read(self._read_from_browser, 10000)
         except BlockingIOError:
@@ -136,7 +101,7 @@ class Pipe:
             if debug:
                 print(f"caught OSError in read() {e!s}", file=sys.stderr)
             if not raw_buffer or raw_buffer == b"{bye}\n":
-                raise PipeClosedError from e
+                raise ChannelClosedError from e
             # this could be hard to test as it is a real OS corner case
         decoded_buffer = raw_buffer.decode("utf-8")
         if debug:
@@ -144,8 +109,8 @@ class Pipe:
         for raw_message in decoded_buffer.split("\0"):
             if raw_message:
                 try:
-                    jsons.append(self.deserialize(raw_message))
-                except simplejson.decoder.JSONDecodeError as e:
+                    jsons.append(wire.deserialize(raw_message))
+                except JSONError as e:
                     if debug:
                         print(
                             f"Problem with {raw_message} in json: {e}",
@@ -160,7 +125,7 @@ class Pipe:
 
     def _unblock_fd(self, fd):
         try:
-            if with_block:
+            if _with_block:
                 os.set_blocking(fd, False)
         except BaseException as e:  # noqa: BLE001 OS errors are not consistent, catch blind
             # also, best effort.
